requirements.txt
fastapi==0.115.8
importlib-metadata==8.0.0
paho-mqtt==2.1.0
platformdirs==4.2.2
uvicorn==0.34.0
pyyaml==6.0.2
opencv-python-headless==4.11.0.86
pillow==11.1.0
python-dateutil==2.9.0.post0
onnxruntime==1.19.2
numpy==1.26.4
requests
websockets
httpx